+++ src/sphere.c
@@ -1,7 +1,7 @@
 #include <math.h>
 #include "sphere.h"
 
-double sphere_hit_distance(sphere_t s, ray_t r) {
+float sphere_hit_distance(sphere_t s, ray_t r) {
 
   point3_t current_center;
   if (!s.moving){
@@ -15,17 +15,17 @@ double sphere_hit_distance(sphere_t s, ray_t r) {
 
   // quadratic equation
 
-  double a            = vec3_dot(r.dir, r.dir);
-  double b            = -2.0 * vec3_dot(r.dir, oc);
-  double c            = vec3_dot(oc, oc) - s.radius * s.radius;
-  double discriminant = b * b - 4 * a * c;
+  float a            = vec3_dot(r.dir, r.dir);
+  float b            = -2.0 * vec3_dot(r.dir, oc);
+  float c            = vec3_dot(oc, oc) - s.radius * s.radius;
+  float discriminant = b * b - 4 * a * c;
 
   if (discriminant < 0) {
     return -1.0; // No valid intersection
   }
 
-  double t0 = (-b - sqrt(discriminant)) / (2.0 * a);
-  double t1 = (-b + sqrt(discriminant)) / (2.0 * a);
+  float t0 = (-b - sqrt(discriminant)) / (2.0 * a);
+  float t1 = (-b + sqrt(discriminant)) / (2.0 * a);
 
   if (t0 > 0) {
     return t0;
