+++ src/sphere.h
@@ -5,6 +5,6 @@
 #include "point3.h"
 #include "ray.h"
 
-double sphere_hit_distance(sphere_t sphere, ray_t ray);
+float sphere_hit_distance(sphere_t sphere, ray_t ray);
 ray_t sphere_center(point3_t center1, point3_t center2);
 #endif // SPHERE_H
